models/team03_PiSAMAP/requirements.txt
diffusers==0.25.0
torch==2.0.1+cu118
transformers==4.28.1
xformers==0.0.20
einops==0.7.0
open-clip-torch==2.20.0
peft==0.9.0
Pillow==9.5.0
PyYAML==6.0
huggingface_hub==0.25.2
numpy==1.23.5
loralib
basicsr
fairscale
pytorch-lightning==1.8.6
tqdm
bcolz-zipline
prettytable
menpo
mxnet
opencv-python
einops
lmdb
pyyaml
scikit-learn
scipy
tensorboard
timm
torchvision
yapf